refactor(interaction): split selection, subchart & zoom
- Refactor resolver
- Change interface: zoom.enabled.type -> zoom.type
