+++ demo/demo.js
@@ -1517,9 +1517,8 @@ setTimeout(function() {chart.resize();}, 3000)
 					}
 				},
 				zoom: {
-					enabled: {
-						type: "drag"
-					}
+					enabled: true,
+					type: "drag"
 				},
 				tooltip: {
 					format: {
