+++ src/config/resolver/axis.ts
@@ -8,68 +8,51 @@
 // Chart
 import apiAxis from "../../Chart/api/axis";
 import apiCategory from "../../Chart/api/category";
-import apiFlow from "../../Chart/api/flow";
 import apiXGrid from "../../Chart/api/grid.x";
 import apiYGrid from "../../Chart/api/grid.y";
 import apiGroup from "../../Chart/api/group";
 import apiRegion from "../../Chart/api/regions";
-import apiSelection from "../../Chart/api/selection";
 import apiX from "../../Chart/api/x";
-import apiZoom from "../../Chart/api/zoom";
+import apiFlow from "../../Chart/api/flow";
 
 // ChartInternal
 import axis from "../../ChartInternal/Axis/Axis";
-import drag from "../../ChartInternal/interactions/drag";
-import flow from "../../ChartInternal/interactions/flow";
 import eventrect from "../../ChartInternal/interactions/eventrect";
-import subchart from "../../ChartInternal/interactions/subchart";
-import zoom from "../../ChartInternal/interactions/zoom";
+import flow from "../../ChartInternal/interactions/flow";
 
 import clip from "../../ChartInternal/internals/clip";
 import grid from "../../ChartInternal/internals/grid";
 import region from "../../ChartInternal/internals/region";
-import selection from "../../ChartInternal/internals/selection";
 import sizeAxis from "../../ChartInternal/internals/size.axis";
 
 // Axis based options
 import optDataAxis from "../Options/data/axis";
 import optAxis from "../Options/axis/axis";
 import optGrid from "../Options/common/grid";
-import optSubchart from "../Options/common/subchart";
-import optZoom from "../Options/common/zoom";
 
 export const api = [
 	apiAxis,
 	apiCategory,
-	apiFlow,
 	apiXGrid,
 	apiYGrid,
+	apiFlow,
 	apiGroup,
 	apiRegion,
-	apiSelection,
-	apiX,
-	apiZoom
+	apiX
 ];
 
 export const internal = [
 	axis,
-	drag,
-	flow,
-	subchart,
-	zoom,
-
 	clip,
 	eventrect,
+	flow,
 	grid,
 	region,
-	selection,
-	sizeAxis
+	sizeAxis,
 ];
 
 export const options = [
 	optDataAxis,
 	optAxis,
 	optGrid,
-	optSubchart,
-	optZoom
 ];
