--- src/config/resolver/index.ts
@@ -1,38 +1,0 @@
-/**
- * Copyright (c) 2017 ~ present NAVER Corp.
- * billboard.js project is licensed under the MIT license
- */
-/**
- * Modules exports
- */
-// Axis
-export {
-	api as axisAPI,
-	internal as axisInternal,
-	options as axisOptions
-} from "./axis";
-
-// Shape
-export {default as shapeArc} from "../../ChartInternal/shape/arc";
-export {default as shapeArea} from "../../ChartInternal/shape/area";
-export {default as shapeBar} from "../../ChartInternal/shape/bar";
-export {default as shapeGauge} from "../../ChartInternal/shape/gauge";
-export {default as shapeBubble} from "../../ChartInternal/shape/bubble";
-export {default as shapeLine} from "../../ChartInternal/shape/line";
-export {default as shapePoint} from "../../ChartInternal/shape/point";
-export {default as shapeRadar} from "../../ChartInternal/shape/radar";
-
-// Options
-export {default as optPoint} from "../Options/common/point";
-export {default as optArea} from "../Options/shape/area";
-export {default as optBar} from "../Options/shape/bar";
-export {default as optBubble} from "../Options/shape/bubble";
-export {default as optLine} from "../Options/shape/line";
-export {default as optScatter} from "../Options/shape/scatter";
-export {default as optSpline} from "../Options/shape/spline";
-
-// Non-Axis based
-export {default as optDonut} from "../Options/shape/donut";
-export {default as optGauge} from "../Options/shape/gauge";
-export {default as optPie} from "../Options/shape/pie";
-export {default as optRadar} from "../Options/shape/radar";
