+++ src/index.ts
@@ -2,19 +2,15 @@
  * Copyright (c) 2017 ~ present NAVER Corp.
  * billboard project is licensed under the MIT license
  */
-import Chart from "./Chart/Chart";
-import ChartInternal from "./ChartInternal/ChartInternal";
-import Options from "./config/Options/Options";
-import {extend} from "./module/util";
-import * as m from "./config/resolver";
+import * as shape from "./config/resolver/shape";
+import * as interaction from "./config/resolver/interaction";
 
-const modules = Object.keys(m);
-const internals = modules.filter(v => /^(axisInternal|shape)/.test(v)).map(v => m[v]);
-const options = modules.filter(v => /^opt/.test(v)).map(v => m[v]);
+// extends shape modules
+Object.keys(shape)
+	.forEach(v => shape[v]());
 
-// extends
-extend(ChartInternal.prototype, internals);
-extend(Chart.prototype, m.axisAPI);
-Options.setOptions(m.axisOptions.concat(options));
+// extends interaction modules
+Object.keys(interaction)
+	.forEach(v => interaction[v]());
 
 export {default, bb} from "./core";
